requirements.txt
groq==0.12.0
fastapi==0.115.6
python-dotenv==1.0.1